make basic setup

+++ product-store/backend/controllers/product.controller.js
@@ -53,7 +53,11 @@ export const updateProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
-
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
   try {
     await Product.findByIdAndDelete(id);
     res
@@ -62,8 +66,6 @@ export const deleteProduct = async (req, res) => {
   } catch (error) {
     console.error("Error in deleting product", error.message);
 
-    res
-      .status(404)
-      .json({ success: false, message: `Product with id: ${id} not found` });
+    res.status(500).json({ success: false, message: `Server Error` });
   }
 };
